frontend/supply-page: доработал таблицу отображения всех записей по поставкам, при пагинации вперед подгружаються новые данные, а при перемотке назад берется из хэшированных данных
