+++ src/Components/Supply/PaginationTable/PaginationTable.tsx
@@ -1,203 +1,23 @@
+import EnhancedTableToolbar from './TableToolbar/TableToolbar';
+import EnhancedTableHead from './TableHead/TableHead';
 import { useGetSuppliesBetweenMutation } from '../../../Store/services/supply';
 import { ISupplies } from '../../../interfaces/ISupply';
 import * as React from 'react';
 import {useEffect} from 'react';
-import { alpha } from '@mui/material/styles';
 import Box from '@mui/material/Box';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
 import TableContainer from '@mui/material/TableContainer';
-import TableHead from '@mui/material/TableHead';
 import TablePagination from '@mui/material/TablePagination';
 import TableRow from '@mui/material/TableRow';
-import TableSortLabel from '@mui/material/TableSortLabel';
-import Toolbar from '@mui/material/Toolbar';
-import Typography from '@mui/material/Typography';
 import Paper from '@mui/material/Paper';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import Switch from '@mui/material/Switch';
-import { visuallyHidden } from '@mui/utils';
 import {DeleteForever} from '@mui/icons-material';
 import { Button } from '@mui/material';
 
-function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
-  if (b[orderBy] < a[orderBy]) {
-    return -1;
-  }
-  if (b[orderBy] > a[orderBy]) {
-    return 1;
-  }
-  return 0;
-}
-
 type Order = 'asc' | 'desc';
-
-function getComparator<Key extends keyof any>(
-  order: Order,
-  orderBy: Key,
-): (
-  a: { [key in Key]: number | string },
-  b: { [key in Key]: number | string },
-) => number {
-  return order === 'desc'
-    ? (a, b) => descendingComparator(a, b, orderBy)
-    : (a, b) => -descendingComparator(a, b, orderBy);
-}
-
-function stableSort<T>(array: readonly T[], comparator: (a: T, b: T) => number) {
-  const stabilizedThis = array.map((el, index) => [el, index] as [T, number]);
-  stabilizedThis.sort((a, b) => {
-    const order = comparator(a[0], b[0]);
-    if (order !== 0) {
-      return order;
-    }
-    return a[1] - b[1];
-  });
-  return stabilizedThis.map((el) => el[0]);
-}
-
-interface HeadCell {
-  disablePadding: boolean;
-  id: keyof ISupplies;
-  label: string;
-  numeric: boolean;
-}
-
-const headCells: readonly HeadCell[] = [
-  {
-    id: 'source',
-    numeric: false,
-    disablePadding: true,
-    label: 'Поставщик',
-  },
-  {
-    id: 'target',
-    numeric: true,
-    disablePadding: false,
-    label: 'Хранилище',
-  },
-  {
-    id: 'item_name',
-    numeric: true,
-    disablePadding: false,
-    label: 'Товар',
-  },
-  {
-    id: 'qty',
-    numeric: true,
-    disablePadding: false,
-    label: 'Количество',
-  },
-  {
-    id: 'price',
-    numeric: true,
-    disablePadding: false,
-    label: 'Цена за единицу',
-  },
-  {
-    id: 'total_price',
-    numeric: true,
-    disablePadding: false,
-    label: 'Общая цена',
-  },
-  {
-    id: 'date',
-    numeric: true,
-    disablePadding: false,
-    label: 'Дата',
-  }
-];
-
-interface EnhancedTableProps {
-  numSelected: number;
-  onRequestSort: (event: React.MouseEvent<unknown>, property: keyof ISupplies) => void;
-  onSelectAllClick: (event: React.ChangeEvent<HTMLInputElement>) => void;
-  order: Order;
-  orderBy: string;
-  rowCount: number;
-}
-
-function EnhancedTableHead(props: EnhancedTableProps) {
-  const { onSelectAllClick, order, orderBy, numSelected, rowCount, onRequestSort } =
-    props;
-  const createSortHandler =
-    (property: keyof ISupplies) => (event: React.MouseEvent<unknown>) => {
-      onRequestSort(event, property);
-    };
-
-  return (
-    <TableHead>
-      <TableRow>
-        <TableCell padding="checkbox">
-          
-        </TableCell>
-        {headCells.map((headCell) => (
-          <TableCell
-            key={headCell.id}
-            align={headCell.numeric ? 'right' : 'left'}
-            padding={headCell.disablePadding ? 'none' : 'normal'}
-            sortDirection={orderBy === headCell.id ? order : false}
-          >
-            <TableSortLabel
-              active={orderBy === headCell.id}
-              direction={orderBy === headCell.id ? order : 'asc'}
-              onClick={createSortHandler(headCell.id)}
-            >
-              {headCell.label}
-              {orderBy === headCell.id ? (
-                <Box component="span" sx={visuallyHidden}>
-                  {order === 'desc' ? 'sorted descending' : 'sorted ascending'}
-                </Box>
-              ) : null}
-            </TableSortLabel>
-          </TableCell>
-        ))}
-      </TableRow>
-    </TableHead>
-  );
-}
-
-interface EnhancedTableToolbarProps {
-  numSelected: number;
-}
-
-function EnhancedTableToolbar(props: EnhancedTableToolbarProps) {
-  const { numSelected } = props;
-
-  return (
-    <Toolbar
-      sx={{
-        pl: { sm: 2 },
-        pr: { xs: 1, sm: 1 },
-        ...(numSelected > 0 && {
-          bgcolor: (theme) =>
-            alpha(theme.palette.primary.main, theme.palette.action.activatedOpacity),
-        }),
-      }}
-    >
-      {numSelected > 0 ? (
-        <Typography
-          sx={{ flex: '1 1 100%' }}
-          color="inherit"
-          variant="subtitle1"
-          component="div"
-        >
-          {numSelected} selected
-        </Typography>
-      ) : (
-        <Typography
-          sx={{ flex: '1 1 100%' }}
-          variant="h6"
-          id="tableTitle"
-          component="div"
-        >
-          Приходы
-        </Typography>
-      )}
-    </Toolbar>
-  );
-}
 
 export default function PaginationTable() {
   const [order, setOrder] = React.useState<Order>('asc');
@@ -209,14 +29,14 @@ export default function PaginationTable() {
   const [rows, setRows] = React.useState<Array<ISupplies> | []>([]);
   const [getSuppliesBetween] = useGetSuppliesBetweenMutation();
 
-  useEffect(() => {
-    initialState();
-  }, [])
-
-  const initialState = async() => {
+  const initialRows = async() => {
     const result = await getSuppliesBetween({start: 1, end:5});
     setRows((result as {data: Array<ISupplies>}).data);
   };
+
+  useEffect(() => {
+    initialRows();
+  }, [])
 
   const handleRequestSort = (
     event: React.MouseEvent<unknown>,
@@ -229,11 +49,9 @@ export default function PaginationTable() {
 
   const handleSelectAllClick = (event: React.ChangeEvent<HTMLInputElement>) => {};
 
-  const handleClick = (event: React.MouseEvent<unknown>, name: string) => {};
-
-  const handleChangePage = async(event: unknown, newPage: number) => {
-    if(newPage > page) {
-      const result = await getSuppliesBetween({start: (page-1)* rowsPerPage + 1 , end: page * rowsPerPage});
+  const handleChangePage = async(event: unknown, newPage: number) => {  
+    if(newPage > page && newPage * rowsPerPage === rows.length) {
+      const result = await getSuppliesBetween({start: (newPage)* rowsPerPage + 1 , end: (newPage+1) * rowsPerPage});
       const resultData = (result as {data: Array<never>}).data;
       setRows(prev => prev.concat(resultData));
     }
@@ -249,18 +67,22 @@ export default function PaginationTable() {
   };
 
   const handleChangeDense = (event: React.ChangeEvent<HTMLInputElement>) => {
-    
     setDense(event.target.checked);
   };
 
-  const onClickDelete = (event: React.ChangeEvent<HTMLButtonElement>, id: number) => {
-
-  };
-
-  const isSelected = (name: string) => selected.indexOf(name) !== -1;
+  const onClickDelete = (event: React.ChangeEvent<HTMLButtonElement>, id: number) => {};
 
   const emptyRows = page > 0 ? Math.max(0, (1 + page) * rowsPerPage - rows.length) : 0;
 
+  const visibleRows = React.useMemo(
+    () =>
+      rows.slice(
+        page * rowsPerPage,
+        page * rowsPerPage + rowsPerPage,
+      ),
+    [rows,page, rowsPerPage]
+  );
+  
   return (
     <Box sx={{ width: '100%' }}>
       <Paper sx={{ width: '100%', mb: 2 }}>
@@ -280,23 +102,17 @@ export default function PaginationTable() {
               rowCount={rows.length}
             />
             <TableBody>
-              {rows.map((row, index) => {
-                const isItemSelected = isSelected(row.source);
+              {rows.length > 0?  visibleRows.map((row, index) => {
                 const labelId = `enhanced-table-checkbox-${index}`;
-
                 return (
                   <TableRow
                     hover
-                    // onClick={(event) => handleClick(event, row.source)}
                     role="checkbox"
-                    aria-checked={isItemSelected}
                     tabIndex={-1}
-                    key={row.source}
-                    selected={isItemSelected}
+                    key={row.id}
                     sx={{ cursor: 'pointer' }}
                   >
-                    <TableCell >
-                    </TableCell>
+                    <TableCell align="left">{row.id}</TableCell>
                     <TableCell
                       component="th"
                       id={labelId}
@@ -324,7 +140,8 @@ export default function PaginationTable() {
                     </TableCell>
                   </TableRow>
                 );
-              })}
+              }): []
+              }
               {emptyRows > 0 && (
                 <TableRow
                   style={{
