+++ src/Components/Supply/PaginationTable/TableToolbar/TableToolbar.tsx
@@ -1,0 +1,43 @@
+import { Toolbar, Typography } from "@mui/material";
+import { alpha } from '@mui/material/styles';
+
+interface EnhancedTableToolbarProps {
+    numSelected: number;
+}
+
+export default function EnhancedTableToolbar(props: EnhancedTableToolbarProps) {
+    const { numSelected } = props;
+  
+    return (
+      <Toolbar
+        sx={{
+          pl: { sm: 2 },
+          pr: { xs: 1, sm: 1 },
+          ...(numSelected > 0 && {
+            bgcolor: (theme) =>
+              alpha(theme.palette.primary.main, theme.palette.action.activatedOpacity),
+          }),
+        }}
+      >
+        {numSelected > 0 ? (
+          <Typography
+            sx={{ flex: '1 1 100%' }}
+            color="inherit"
+            variant="subtitle1"
+            component="div"
+          >
+            {numSelected} selected
+          </Typography>
+        ) : (
+          <Typography
+            sx={{ flex: '1 1 100%' }}
+            variant="h6"
+            id="tableTitle"
+            component="div"
+          >
+            Приходы
+          </Typography>
+        )}
+      </Toolbar>
+    );
+}
